Break piece moving methods into a concern
This remedies the Rubocop gripe that the piece model was too long
(over 100 lines).
